% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnabatTools.R
\name{calcFreq}
\alias{calcFreq}
\title{Internal function to compute frequency values}
\usage{
calcFreq(params, timeData, N)
}
\arguments{
\item{params}{Parameter set as produced by function \code{\link{getParams}}}

\item{timeData}{Named list produced by either \code{\link{getData129}} or \code{\link{getData130}}}

\item{N}{Index (integer) to the last item in the timeData data set}
}
\value{
Named list: freq, showDot
}
\description{
Internal function to compute frequency values
}
\seealso{
\code{\link{getParams}}, \code{\link{getData129}}, \code{\link{getData130}}
}

