% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnabatTools.R
\name{callStats}
\alias{callStats}
\title{Compute statistics from data in an Anabat object}
\usage{
callStats(thisObj, timeSpan = "full", timeLimits = NULL,
  cleanBelow = NULL)
}
\arguments{
\item{thisObj}{An object of class Anabat produced by the function \code{\link{readAnabat}}}

\item{timeSpan}{Character value indicating if statistics are to be computed on the "full" data sequence or a "partial" data sequence. Default = "full"}

\item{timeLimits}{If a partial sequence, then this 2-element numeric vector gives the start time (in seconds) and the end time (in seconds) used to "window" the sequence}

\item{cleanBelow}{A frequency filter value. All points less than or equal to the value (in Hertz) will be cleaned from the data before statistics are computed. A value of NULL (default) indicates no frequency filtering}
}
\value{
A named list with basic call statistics \emph{for the selected window}:
\tabular{ll}{
  meanCallDuration \tab Mean duration of a call (in seconds)\cr
  meanCallRate\tab Mean number of calls per second\cr
  meanCallGap\tab Mean time between calls (in seconds)\cr
  freqRange\tab Two-element numeric vector holding Minimum and Maximum frequencies (Hz)\cr
  meanFreq\tab Mean frequency (Hz)\cr
  medianFreq\tab Median frequency (Hz)\cr
  }
}
\description{
Computes some basic statistics on the data supplied in parameter thisObj. Data can be restricuted to a rectangular window by specifying a start and end time, and independently, a lower frequency bound. Statistics are only computed on data remaining within the window. By default the entire data set is analysed.
}
\examples{
d <- readAnabat(system.file("extdata","Miau-ne.01#",package="AnabatTools"))

# Basic stats for whole data set:
stats1 <- callStats(d)

# Which is the same as the raw statistics contained in the object created by readAnabat:

# Window on time axis:
stats2 <- callStats(d, startTime = 2.4, endTime = 4)

# Window on the frequency axis using cleanBelow:
stats3 <- callStats(d, cleanBelow = 50000)

# Combined window:
stats3 <- callStats(d, startTime = 2.4, endTime = 4, cleanBelow = 50000)
}

