% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnabatTools.R
\name{getData129}
\alias{getData129}
\title{Internal function to extract time-frequency data from a binary Anabat file for Filetype 129}
\usage{
getData129(p, params, theData)
}
\arguments{
\item{p}{Pointer (integer) to the start of the block containing call data}

\item{params}{Parameter set as produced by function getParams}

\item{theData}{A binary Anabat data file. These data are further processed by function \code{\link{calcFreq}} to derive frequency data}
}
\value{
A named list: timeData, last.t, and showDot
}
\description{
Internal function to extract time-frequency data from a binary Anabat file for Filetype 129
}

