% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnabatTools.R
\name{getData130}
\alias{getData130}
\title{Internal function to extract time-frequency data from a binary Anabat file for Filetypes 130, 131 and 132}
\usage{
getData130(p, params, fileType, theData)
}
\arguments{
\item{p}{Pointer (integer) to the start of the block containing call data}

\item{params}{Parameter set as produced by function getParams}

\item{fileType}{Integer indicating the file type: 130, 131 or 132}

\item{theData}{A binary Anabat data file}
}
\value{
A named list: timeData, last.t, and showDot. These data are further processed by function \code{\link{calcFreq}} to derive frequency data
}
\description{
Internal function to extract time-frequency data from a binary Anabat file for Filetypes 130, 131 and 132
}

