% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnabatTools.R
\name{getDateTime}
\alias{getDateTime}
\title{Internal function to extract Date and Time from a Type 132 binary Anabat file}
\usage{
getDateTime(fileType, theData)
}
\arguments{
\item{fileType}{Integer indicating Anabat file type}

\item{theData}{A binary Anabat data file}

\item{p}{Pointer (integer) to start of identified byte string containing encoded date and time}
}
\value{
A named list with the following integer-valued elements: DAY, MONTH, YEAR, HOURS, MINS, SECS, HUNDS, and MICROS
}
\description{
Internal function to extract Date and Time from a Type 132 binary Anabat file
}

