% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnabatTools.R
\name{getParams}
\alias{getParams}
\title{Internal function to extract key paramaters from a binary Anabat file}
\usage{
getParams(p, theData)
}
\arguments{
\item{p}{Pointer (integer) to the start of the byte string containing paramaters to be extracted}

\item{theData}{A binary Anabat data file}
}
\value{
A named list with the following elements: RES1, DIVRAT, VRES, and timeFactor
}
\description{
Internal function to extract key paramaters from a binary Anabat file
}

