% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnabatTools.R
\name{getTextHeader}
\alias{getTextHeader}
\title{Internal function to extract text header information from a binary Anabat file}
\usage{
getTextHeader(theData)
}
\arguments{
\item{theData}{A binary data file.}
}
\value{
A named list with elements holding extracted fields from the text header. Fields are: TAPE, DATE, LOC, SPECIES, SPEC, NOTE and NOTE1.
}
\description{
Internal function to extract text header information from a binary Anabat file
}

