% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnabatTools.R
\name{plot.Anabat}
\alias{plot.Anabat}
\title{Generic plot function for Anabat objects}
\usage{
\method{plot}{Anabat}(AnabatData, showLines = T, showLineType = 2,
  lineColour = "grey50", maxFreq = 125000, minFreq = 0,
  cleanBelow = NULL, startTime = NULL, endTime = NULL, dotPch = 16,
  dotCex = 0.3, dotCol = "blue")
}
\arguments{
\item{AnabatData}{An object of class Anabat as produced by the function readAnabat}

\item{showLines}{Logical flag to indicate if horizontal grid lines should be produced on the plot. Default is TRUE.}

\item{showLineType}{Integer (default = 2, "dashed") indicating the line type to be used for horizontal grid lines. See help for par and option "lty"}

\item{lineColour}{Colour to be used for horizontal grid lines. This can be any valid value indexing a colour from the currently active palette.}

\item{maxFreq}{Maximum frequency to be used for the plot expressed in Hertz. That is, 35kHz would be entered as 30000. Default = 125000.}

\item{minFreq}{Minimum frequency to be used for the plot expressed in Hertz. That is, 5kHz would be entered as 5000. Default = 0.}

\item{cleanBelow}{Apply a simple frequency filter so that any points with frequency less than cleanBelow (in Hertz as described above) will be omitted from the plot.}

\item{startTime}{Time in seconds for the left margin of the plot. Default = 0 (ie start of the data sequence).}

\item{endTime}{Time in seconds for the right margin of the plot. Default = maximum time point in the data (ie end of the data sequence).}

\item{dotPch}{Numeric or character value specifying the plot symbol. See help for points function in the base graphics package for options. Default is ".".}

\item{dotCex}{Size of the specified plot symbol as a decimal fraction of the standard character size. Default = 0.3.}

\item{dotCol}{Colour to be used for the plot symbol. This can be any valid value (integer or character) indexing a colour from the currently active palette. Default = "blue" from the default R palette.}
}
\description{
Generic plot function for Anabat objects
}
\examples{
d <- readAnabat(system.file("extdata","Miau-ne.01#",package="AnabatTools"))

# Basic plot:
plot(d)

# Change plot colour using a named colour from the default R palette:
plot(d, dotCol = "tomato")

# Window on time axis:
plot(d, startTime = 2.4, endTime = 4)

# Window on the frequency axis using cleanBelow:
plot(d, cleanBelow = 50000)

# Combined window:
plot(d, startTime = 2.4, endTime = 4, cleanBelow = 50000)
}

