% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnabatTools.R
\name{readAnabat}
\alias{readAnabat}
\title{Read and parse an Anabat zero-crossing data file returning an S3 object containing the recovered data}
\usage{
readAnabat(fileName)
}
\arguments{
\item{fileName}{Path to an Anabat data file}
}
\value{
An S3 object containing the data extracted from fileName
}
\description{
The workhorse function which reads and parses an Anabat zero-crossing data file and creates an S3 object to contain it. When a valid data file is loaded, the function also calls \code{\link{callStats}} to compute call statistics for the whole or raw data. These results are incorporated into the returned Anabat object.
}
\examples{
d <- readAnabat(system.file("extdata","Miau-ne.01#",package="AnabatTools"))

# Some exploration and quality checking:
summary(d)
plot(d)
}

