% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnabatTools.R
\name{summary.Anabat}
\alias{summary.Anabat}
\title{Generic function to produce a summary of an Anabat object on the console}
\usage{
\method{summary}{Anabat}(AnabatData)
}
\arguments{
\item{AnabatData}{An object of class Anabat as produced by the function readAnabat}
}
\description{
Generic function to produce a summary of an Anabat object on the console
}
\examples{
d <- readAnabat("blah")
summary(d)
}

